src/services/requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
pypdf
chromadb
openai
python-dotenv
